Ensure particle position and velocity is float64
This fixes the behavior of hdf5 trajectories which store float32
coordinates and this is not compatible with some backends.

+++ atooms/system/particle.py
@@ -23,11 +23,11 @@ class Particle(object):
         if position is None:
             self.position = numpy.zeros(_ndim)
         else:
-            self.position = numpy.asarray(position)
+            self.position = numpy.asarray(position, dtype='float64')
         if velocity is None:
             self.velocity = numpy.zeros(_ndim)
         else:
-            self.velocity = numpy.asarray(velocity)
+            self.velocity = numpy.asarray(velocity, dtype='float64')
 
     @property
     def diameter(self):
